read the descripiton
i only did this to try to complete the bar in insights.

however after reading it a bit, i read that any unproffiensial and unwelcome or innaporopitate language (my spelling for that broke the code of conduct already) is it a bit too harsh. i can easily edit it sure but i'd rather delete it.